+++ src/Sulu/Bundle/ContentBundle/Resources/phpcr-migrations/Version201511240843.php
@@ -229,7 +229,7 @@ class Version201511240843 implements VersionInterface, ContainerAwareInterface
             }
         }
 
-        $this->documentManager->persist($document, $locale);
+        $this->documentManager->persist($document, $locale, ['auto_name' => false]);
     }
 
     /**