+++ CHANGELOG.md
@@ -5,6 +5,9 @@ CHANGELOG for Sulu
     * BUGFIX      #1893 [ContentBundle]   Fixed resource locator deferred for edit
     * BUGIFX      #1871 [ContentBundle]   Fixed url-generation and save button
 
+* dev-master
+    * HOTFIX      #1888 [ContentBundle]   used auto_name in phpcr migrations
+
 * 1.1.2 (2015-12-11)
     * HOTFIX      #1831 [MediaBundle]     Fixed query for retrieving entities to index
     * HOTFIX      #1868 [ContentBundle]   Added date upgrade script for blocks